Move deprecated functions to deprecated directory (#643)
