+++ galsim/correlatednoise.py
@@ -24,8 +24,6 @@ import galsim
 from . import base
 from . import utilities
 
-from .deprecated import depr
-
 def whitenNoise(self, noise):
     # This will be inserted into the Image class as a method.  So self = image.
     """Whiten the noise in the image assuming that the noise currently in the image can be described
@@ -281,11 +279,6 @@ class _BaseCorrelatedNoise(galsim.BaseNoise):
         # Return the variance to the interested user
         return variance
 
-    def applyWhiteningTo(self, image):
-        """A deprecated synonym for whitenImage"""
-        depr('applyWhiteningTo', 1.2, 'whitenImage')
-        return self.whitenImage(image)
-
     def symmetrizeImage(self, image, order=4):
         """Apply noise designed to impose N-fold symmetry on the existing noise in a (square) input
         Image.
@@ -387,19 +380,6 @@ class _BaseCorrelatedNoise(galsim.BaseNoise):
         """
         return _BaseCorrelatedNoise(self.getRNG(), self._profile.expand(scale))
 
-    def createExpanded(self, scale):
-        """A deprecated synonym for expand(scale)"""
-        depr('createExpanded', 1.1, 'obj.expand(scale)')
-        return self.expand(scale)
-
-    def applyExpansion(self, scale):
-        """A deprecated method that is roughly equivalent to obj = obj.expand(scale)"""
-        depr('applyExpansion', 1.1, 'obj = obj.expand(scale)')
-        new_obj = self.expand(scale)
-        self._profile = new_obj._profile
-        self._profile_for_stored = None  # Reset the stored profile as it is no longer up-to-date
-        self.__class__ = new_obj.__class__
-
     def dilate(self, scale):
         """Apply the appropriate changes to the scale and variance for when the object has
         an applied dilation.
@@ -412,19 +392,6 @@ class _BaseCorrelatedNoise(galsim.BaseNoise):
         # so the variance needs to change by scale**-4.
         return _BaseCorrelatedNoise(self.getRNG(), self._profile.expand(scale) * (1./scale**4))
 
-    def createDilated(self, scale):
-        """A deprecated synonym for dilate(scale)"""
-        depr('createDilated', 1.1, 'obj.dilate(scale)')
-        return self.dilate(scale)
-
-    def applyDilation(self, scale):
-        """A deprecated method that is roughly equivalent to obj = obj.dilate(scale)"""
-        depr('applyDilation', 1.1, 'obj = obj.dilate(scale)')
-        new_obj = self.dilate(scale)
-        self._profile = new_obj._profile
-        self._profile_for_stored = None  # Reset the stored profile as it is no longer up-to-date
-        self.__class__ = new_obj.__class__
-
     def magnify(self, mu):
         """Apply the appropriate changes to the scale and variance for when the object has
         an applied magnification `mu`.
@@ -434,19 +401,6 @@ class _BaseCorrelatedNoise(galsim.BaseNoise):
         @returns a new CorrelatedNoise object with the specified magnification.
         """
         return _BaseCorrelatedNoise(self.getRNG(), self._profile.magnify(mu))
-
-    def createMagnified(self, mu):
-        """A deprecated synonym for magnify(mu)"""
-        depr('createMagnified', 1.1, 'obj.magnify(mu)')
-        return self.magnify(mu)
-
-    def applyMagnification(self, mu):
-        """A deprecated method that is roughly equivalent to obj = obj.magnify(mu)"""
-        depr('applyMagnification', 1.1, 'obj = obj.magnify(mu)')
-        new_obj = self.magnify(mu)
-        self._profile = new_obj._profile
-        self._profile_for_stored = None  # Reset the stored profile as it is no longer up-to-date
-        self.__class__ = new_obj.__class__
 
     def lens(self, g1, g2, mu):
         """Apply the appropriate changes for when the object has an applied shear and magnification.
@@ -459,19 +413,6 @@ class _BaseCorrelatedNoise(galsim.BaseNoise):
         """
         return _BaseCorrelatedNoise(self.getRNG(), self._profile.lens(g1,g2,mu))
 
-    def createLensed(self, g1, g2, mu):
-        """A deprecated synonym for lens(g1,g2,mu)"""
-        depr('createLensed', 1.1, 'obj.lens(g1,g2,mu)')
-        return self.lens(g1,g2,mu)
-
-    def applyLensing(self, g1, g2, mu):
-        """A deprecated method that is roughly equivalent to obj = obj.lens(g1,g2,mu)"""
-        depr('applyLensing', 1.1, 'obj = obj.lens(g1,g2,mu)')
-        new_obj = self.lens(g1,g2,mu)
-        self._profile = new_obj._profile
-        self._profile_for_stored = None  # Reset the stored profile as it is no longer up-to-date
-        self.__class__ = new_obj.__class__
-
     def rotate(self, theta):
         """Apply a rotation `theta` to this correlated noise model.
 
@@ -482,19 +423,6 @@ class _BaseCorrelatedNoise(galsim.BaseNoise):
         if not isinstance(theta, galsim.Angle):
             raise TypeError("Input theta should be an Angle")
         return _BaseCorrelatedNoise(self.getRNG(), self._profile.rotate(theta))
-
-    def createRotated(self, theta):
-        """A deprecated synonym for rotate(theta)"""
-        depr('createRotated', 1.1, 'obj.rotate(theta)')
-        return self.rotate(theta)
-
-    def applyRotation(self, theta):
-        """A deprecated method that is roughly equivalent to obj = obj.rotate(theta)"""
-        depr('applyRotation', 1.1, 'obj = obj.rotate(theta)')
-        new_obj = self.rotate(theta)
-        self._profile = new_obj._profile
-        self._profile_for_stored = None  # Reset the stored profile as it is no longer up-to-date
-        self.__class__ = new_obj.__class__
 
     def shear(self, *args, **kwargs):
         """Apply a shear to this correlated noise model, where arguments are either a Shear,
@@ -510,19 +438,6 @@ class _BaseCorrelatedNoise(galsim.BaseNoise):
         """
         return _BaseCorrelatedNoise(self.getRNG(), self._profile.shear(*args,**kwargs))
 
-    def createSheared(self, *args, **kwargs):
-        """A deprecated synonym for shear(shear)"""
-        depr('createSheared', 1.1, 'obj.shear(shear)')
-        return self.shear(*args,**kwargs)
-
-    def applyShear(self, *args, **kwargs):
-        """A deprecated method that is roughly equivalent to obj = obj.shear(shear)"""
-        depr('applyShear', 1.1, 'obj = obj.shear(shear)')
-        new_obj = self.shear(*args, **kwargs)
-        self._profile = new_obj._profile
-        self._profile_for_stored = None  # Reset the stored profile as it is no longer up-to-date
-        self.__class__ = new_obj.__class__
-
     def transform(self, dudx, dudy, dvdx, dvdy):
         """Apply an arbitrary jacobian transformation to this correlated noise model.
 
@@ -534,19 +449,6 @@ class _BaseCorrelatedNoise(galsim.BaseNoise):
         @returns a new CorrelatedNoise object with the specified transformation.
         """
         return _BaseCorrelatedNoise(self.getRNG(), self._profile.transform(dudx,dudy,dvdx,dvdy))
-
-    def createTransformed(self, dudx, dudy, dvdx, dvdy):
-        """A deprecated synonym for transform(dudx,dudy,dvdx,dvdy)"""
-        depr('createTransformed', 1.1, 'obj.transform(dudx,dudy,dvdx,dvdy)')
-        return self.transform(dudx,dudy,dvdx,dvdy)
-
-    def applyTransformation(self, dudx, dudy, dvdx, dvdy):
-        """A deprecated method that is roughly equivalent to obj = obj.transform(...)"""
-        depr('applyTransformation', 1.1, 'obj = obj.transform(dudx,dudy,dvdx,dvdy)')
-        new_obj = self.transform(dudx,dudy,dvdx,dvdy)
-        self._profile = new_obj._profile
-        self._profile_for_stored = None  # Reset the stored profile as it is no longer up-to-date
-        self.__class__ = new_obj.__class__
 
     def getVariance(self):
         """Return the point variance of this noise field, equal to its correlation function value at
@@ -603,24 +505,6 @@ class _BaseCorrelatedNoise(galsim.BaseNoise):
         """
         return _BaseCorrelatedNoise(self.getRNG(), self._profile * variance_ratio)
 
-    def setVariance(self, variance):
-        """A deprecated method that is roughly equivalent to
-        corr = corr.withVariance(variance)
-        """
-        depr('setVariance', 1.1, 'obj = obj.withVariance(variance)')
-        new_obj = self.withVariance(variance)
-        self._profile = new_obj._profile
-        self._profile_for_stored = None  # Reset the stored profile as it is no longer up-to-date
-        self.__class__ = new_obj.__class__
-
-    def scaleVariance(self, variance_ratio):
-        """A deprecated method that is roughly equivalent to corr = corr * variance_ratio"""
-        depr('scaleVariance', 1.1, 'obj = obj * variance_ratio')
-        new_obj = self.withScaledVariance(variance_ratio)
-        self._profile = new_obj._profile
-        self._profile_for_stored = None  # Reset the stored profile as it is no longer up-to-date
-        self.__class__ = new_obj.__class__
-
     def convolvedWith(self, gsobject, gsparams=None):
         """Convolve the correlated noise model with an input GSObject.
 
@@ -671,16 +555,6 @@ class _BaseCorrelatedNoise(galsim.BaseNoise):
         conv = galsim.Convolve([self._profile, galsim.AutoCorrelate(gsobject)], gsparams=gsparams)
         return _BaseCorrelatedNoise(self.getRNG(), conv)
 
-    def convolveWith(self, gsobject, gsparams=None):
-        """A deprecated method that is roughly equivalent to
-        cn = cn.convolvedWith(gsobject,gsparams)
-        """
-        depr('convolveWith', 1.1, 'obj = obj.convolvedWith(gsobject, gsparams)')
-        new_obj = self.convolvedWith(gsobject,gsparams)
-        self._profile = new_obj._profile
-        self._profile_for_stored = None  # Reset the stored profile as it is no longer up-to-date
-        self.__class__ = new_obj.__class__
-
     def drawImage(self, image=None, scale=None, dtype=None, wmult=1., add_to_image=False, dx=None):
         """A method for drawing profiles storing correlation functions.
 
@@ -714,34 +588,13 @@ class _BaseCorrelatedNoise(galsim.BaseNoise):
         @returns an Image of the correlation function.
         """
         # Check for obsolete dx parameter
+        from galsim.deprecated import depr
         depr('dx', 1.1, 'scale')
         if dx is not None and scale is None: scale = dx
 
         return self._profile.drawImage(
             image=image, scale=scale, dtype=dtype, method='sb', gain=1., wmult=wmult,
             add_to_image=add_to_image, use_true_center=False)
-
-    def draw(self, *args, **kwargs):
-        """A deprecated synonym of drawImage"""
-        depr('draw', 1.1, "drawImage")
-        return self.drawImage(*args, **kwargs)
-
-    def calculateCovarianceMatrix(self, bounds, scale):
-        """Calculate the covariance matrix for an image with specified properties.
-
-        A correlation function also specifies a covariance matrix for noise in an image of known
-        dimensions and pixel scale.  The user specifies these bounds and pixel scale, and this
-        method returns a covariance matrix as a square Image object, with the upper triangle
-        containing the covariance values.
-
-        @param  bounds Bounds corresponding to the dimensions of the image for which a covariance
-                       matrix is required.
-        @param  scale  Pixel scale of the image for which a covariance matrix is required.
-
-        @returns the covariance matrix (as an Image).
-        """
-        # TODO: Allow this to take a JacobianWCS, rather than just a scale.
-        return galsim._galsim._calculateCovarianceMatrix(self._profile.SBProfile, bounds, scale)
 
     def _get_update_rootps(self, shape, wcs):
         """Internal utility function for querying the `rootps` cache, used by applyTo(),
@@ -1200,6 +1053,7 @@ class CorrelatedNoise(_BaseCorrelatedNoise):
     def __init__(self, image, rng=None, scale=0., x_interpolant=None, correct_periodicity=True,
         subtract_mean=False, gsparams=None, dx=None):
         # Check for obsolete dx parameter
+        from galsim.deprecated import depr
         depr('dx', 1.1, 'scale')
         if dx is not None and scale==0.: scale = dx
 
@@ -1321,17 +1175,6 @@ def _cf_periodicity_dilution_correction(cf_shape):
     return correction
 
 
-# Make a function for returning Noise correlations
-def _Image_getCorrelatedNoise(image):
-    """Deprecated method to get a CorrelatedNoise instance by calculating the correlation function
-    of image pixels.  It is equivalent to `noise = galsim.CorrelatedNoise(image)`
-    """
-    depr('getCorrelatedNoise',1.1,'noise = galsim.CorrelatedNoise(image)')
-    return CorrelatedNoise(image)
-
-# Then add this Image method to the Image class
-galsim.Image.getCorrelatedNoise = _Image_getCorrelatedNoise
-
 # Free function for returning a COSMOS noise field correlation function
 def getCOSMOSNoise(file_name=None, rng=None, cosmos_scale=0.03, variance=0., x_interpolant=None,
                    gsparams=None, dx_cosmos=None):
@@ -1427,6 +1270,7 @@ def getCOSMOSNoise(file_name=None, rng=None, cosmos_scale=0.03, variance=0., x_i
     The FITS file `out.fits` should then contain an image of randomly-generated, COSMOS-like noise.
     """
     # Check for obsolete dx_cosmos parameter
+    from galsim.deprecated import depr
     depr('dx_cosmos', 1.1, 'cosmos_scale')
     if dx_cosmos is not None and cosmos_scale==0.03: cosmos_scale = dx_cosmos
